chore(react): switch from vue to react
* there are 3 large components that will need reactive state so best keep with react
