+++ astro.config.mjs
@@ -1,6 +1,6 @@
 import { defineConfig } from "astro/config";
+import react from "@astrojs/react";
 import tailwind from "@astrojs/tailwind";
-import vue from "@astrojs/vue";
 
 // https://astro.build/config
 export default defineConfig({
@@ -8,6 +8,6 @@ export default defineConfig({
     tailwind({
       applyBaseStyles: false,
     }),
-    vue(),
+    react(),
   ],
 });
