+++ package.json
@@ -15,13 +15,14 @@
   },
   "dependencies": {
     "@astrojs/check": "^0.8.2",
+    "@astrojs/react": "^3.6.0",
     "@astrojs/tailwind": "^5.1.0",
-    "@astrojs/vue": "^4.5.0",
     "astro": "^4.12.2",
     "leaflet": "^1.9.4",
     "leaflet.fullscreen": "^3.0.2",
-    "tailwindcss": "^3.4.6",
-    "vue": "^3.4.33"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwindcss": "^3.4.6"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.3.0",
@@ -34,6 +35,8 @@
     "@types/leaflet": "^1.9.12",
     "@types/leaflet.fullscreen": "^3.0.2",
     "@types/node": "^20.14.11",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "@typescript-eslint/parser": "^7.17.0",
     "commitizen": "^4.3.0",
     "cz-conventional-changelog": "^3.3.0",
