+++ tsconfig.json
@@ -5,6 +5,7 @@
     "paths": {
       "~/*": ["./src/*"]
     },
-    "jsx": "preserve"
+    "jsx": "react-jsx",
+    "jsxImportSource": "react"
   }
 }
